test: improve performance
