+++ packages/plugin-commands-env/test/env.test.ts
@@ -49,155 +49,6 @@ test('install Node (and npm, npx) by exact version of Node.js', async () => {
   }
 })
 
-test('install Node by version range', async () => {
-  tempDir()
-
-  await env.handler({
-    bin: process.cwd(),
-    global: true,
-    pnpmHomeDir: process.cwd(),
-    rawConfig: {},
-  }, ['use', '6'])
-
-  const { stdout } = execa.sync('node', ['-v'], {
-    env: {
-      [PATH]: `${process.cwd()}${path.delimiter}${process.env[PATH] as string}`,
-    },
-  })
-  expect(stdout.toString()).toBe('v6.17.1')
-
-  const dirs = fs.readdirSync(path.resolve('nodejs'))
-  expect(dirs).toEqual(['6.17.1'])
-})
-
-test('install the LTS version of Node', async () => {
-  tempDir()
-
-  await env.handler({
-    bin: process.cwd(),
-    global: true,
-    pnpmHomeDir: process.cwd(),
-    rawConfig: {},
-  }, ['use', 'lts'])
-
-  const { stdout: version } = execa.sync('node', ['-v'], {
-    env: {
-      [PATH]: `${process.cwd()}${path.delimiter}${process.env[PATH] as string}`,
-    },
-  })
-  expect(version).toBeTruthy()
-
-  const dirs = fs.readdirSync(path.resolve('nodejs'))
-  expect(dirs).toEqual([version.substring(1)])
-})
-
-test('install Node by its LTS name', async () => {
-  tempDir()
-
-  await env.handler({
-    bin: process.cwd(),
-    global: true,
-    pnpmHomeDir: process.cwd(),
-    rawConfig: {},
-  }, ['use', 'argon'])
-
-  const { stdout: version } = execa.sync('node', ['-v'], {
-    env: {
-      [PATH]: `${process.cwd()}${path.delimiter}${process.env[PATH] as string}`,
-    },
-  })
-  expect(version).toBe('v4.9.1')
-
-  const dirs = fs.readdirSync(path.resolve('nodejs'))
-  expect(dirs).toEqual([version.substring(1)])
-})
-
-test('install the latest Node.js', async () => {
-  tempDir()
-
-  await env.handler({
-    bin: process.cwd(),
-    global: true,
-    pnpmHomeDir: process.cwd(),
-    rawConfig: {},
-  }, ['use', 'latest'])
-
-  const { stdout } = execa.sync('node', ['-v'], {
-    env: {
-      [PATH]: `${process.cwd()}${path.delimiter}${process.env[PATH] as string}`,
-    },
-  })
-  expect(stdout.toString()).toMatch(/^v/)
-})
-
-test('install the latest nightly Node.js', async () => {
-  tempDir()
-
-  await env.handler({
-    bin: process.cwd(),
-    global: true,
-    pnpmHomeDir: process.cwd(),
-    rawConfig: {},
-  }, ['use', 'nightly'])
-
-  const { stdout } = execa.sync('node', ['-v'], {
-    env: {
-      [PATH]: `${process.cwd()}${path.delimiter}${process.env[PATH] as string}`,
-    },
-  })
-  expect(stdout.toString()).toMatch(/^v/)
-})
-
-test('install Node by exact version of rc build', async () => {
-  tempDir()
-  const configDir = path.resolve('config')
-
-  await env.handler({
-    bin: process.cwd(),
-    configDir,
-    global: true,
-    pnpmHomeDir: process.cwd(),
-    rawConfig: {},
-  }, ['use', '16.0.0-rc.0'])
-
-  const opts = {
-    env: {
-      [PATH]: `${process.cwd()}${path.delimiter}${process.env[PATH] as string}`,
-    },
-    extendEnv: false,
-  }
-
-  {
-    const { stdout } = execa.sync('node', ['-v'], opts)
-    expect(stdout.toString()).toBe('v16.0.0-rc.0')
-  }
-})
-
-test('install Node by major version from the specified release directory', async () => {
-  tempDir()
-  const configDir = path.resolve('config')
-
-  await env.handler({
-    bin: process.cwd(),
-    configDir,
-    global: true,
-    pnpmHomeDir: process.cwd(),
-    rawConfig: {},
-  }, ['use', 'rc/10'])
-
-  const opts = {
-    env: {
-      [PATH]: `${process.cwd()}${path.delimiter}${process.env[PATH] as string}`,
-    },
-    extendEnv: false,
-  }
-
-  {
-    const { stdout } = execa.sync('node', ['-v'], opts)
-    expect(stdout.toString()).toBe('v10.23.0-rc.0')
-  }
-})
-
 test('fail if a non-existend Node.js version is tried to be installed', async () => {
   tempDir()
